Create linearAlgebraGUI
This project is all about linear algebra and solving it using coding. The factor used in this is ranking system, Gaussian elimination and matrices and determinants.
It takes all the criteria for forming a linear equations of multivariable and produce a specific answer to it. If the linear equation produces a unique solution it then shows a 3D graph plotting the essential points for the equation to form a plane